+++ packages/angular/src/public-api.ts
@@ -23,7 +23,6 @@ export type {
 
   // Column definition
   ColumnDefinition,
-  ColumnId,
   ColumnState,
   ColumnStateUpdate,
   ColumnStateSnapshot,